Move code from __init__.py to analyzer.py
